refactor(arm64): deduplicate lowering paths
